Refactor: svg 경로 수정
