+++ src/containers/main/Intro.tsx
@@ -1,5 +1,5 @@
-import IntroDuck from '@/../public/svgs/intro/introDuck.svg'
-import Title from '@/../public/svgs/intro/title.svg'
+import IntroDuck from '@/public/svgs/intro/introDuck.svg'
+import Title from '@/public/svgs/intro/title.svg'
 
 export default function Intro() {
   return (
