+++ src/containers/main/LiveAndHotItem.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import exImg from '@/dummydata/banner1.png'
+import LikeFalse from '@/public/svgs/icon/likeFalse.svg'
 import { LiveAndHotType } from '@/types/LiveAndHotType'
-import LikeFalse from '@/../public/svgs/icon/likeFalse.svg'
 
 export default function LiveAndHotItem({ item }: { item: LiveAndHotType }) {
   return (
